Level meter 02: the rename to the house file layout
src/processor.ts -> src/worklet.ts (hand-written) and src/_processor.ts ->
src/processor.ts (generated bundle), matching all 22 other packages, and the
level-meter special case comes out of .github/workflows/test.yml. Adds the
"test": "jest" script the package never had, and a comment on why index.ts
hand-rolls new AudioWorkletNode instead of using createWorkletConstructor.

No behaviour change; the tests land in the next commit.
