--- packages/level-meter/src/_processor.ts
@@ -1,1 +1,0 @@
-export const PROCESSOR = `"use strict";(()=>{var n=class extends AudioWorkletProcessor{peaks;max;r;constructor(e){super(),this.r=!0;let s=e.processorOptions.peaksBuffer;this.peaks=new Float32Array(s),this.max=8,this.port.onmessage=c=>{switch(c.data.type){case"DISPOSE":this.r=!1;break}}}process(e,s,c){let i=e[0],h=s[0],l=Math.min(i.length,this.max);for(let t=0;t<l;t++){let r=i[t],p=h[t],a=0;for(let o=0;o<r.length;o++)a=Math.max(a,Math.abs(r[o]));this.peaks[t]=this.peaks[t]*.9+a*.1,p.set(r)}return this.r}static get parameterDescriptors(){return[]}};registerProcessor("LevelMeterProcessor",n);})();`;
